refactor(Utils, LottoTicket): 6개의 1-45 범위 숫자를 랜덤으로 뽑는 기능 리팩토링
* 유틸 클래스가 로또를 뽑는 상황에 종속되지 않아야한다고 생각해 로또 관련 상수를 LottoTicket 에 옮겼다.
* Utils 클래스를 Util 패키지로 옮겼다.
