--- src/main/java/lotto/Utils.java
@@ -1,24 +1,0 @@
-package lotto;
-
-import camp.nextstep.edu.missionutils.Randoms;
-
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
-
-public class Utils {
-    private static final int MINIMUM_LOTTO_NUMBER = 1;
-    private static final int MAXIMUM_LOTTO_NUMBER = 45;
-    private static final int NUMBER_OF_LOTTO = 6;
-    private Utils() {
-    }
-
-    public static List<Integer> generateRandomLottoNumbers() {
-        List<Integer> lottoNumbers = Randoms.pickUniqueNumbersInRange(MINIMUM_LOTTO_NUMBER, MAXIMUM_LOTTO_NUMBER, NUMBER_OF_LOTTO);
-        lottoNumbers = lottoNumbers.stream()
-                .sorted()
-                .toList();
-        return lottoNumbers;
-    }
-
-}
