+++ src/main/java/lotto/domain/LottoTicket.java
@@ -1,13 +1,16 @@
 package lotto.domain;
 
 import lotto.Lotto;
-import lotto.Utils;
+import lotto.util.Utils;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class LottoTicket {
     private static final int TICKET_PRICE = 1000;
+    private static final int MINIMUM_LOTTO_NUMBER = 1;
+    private static final int MAXIMUM_LOTTO_NUMBER = 45;
+    private static final int NUMBER_OF_LOTTO = 6;
     int ticketNumber;
 
     public LottoTicket(int input) {
@@ -18,7 +21,7 @@ public class LottoTicket {
     public List<Lotto> generateLottos() {
         List<Lotto> lottos = new ArrayList<>();
         for (int ticket = 0; ticket < ticketNumber; ticket++) {
-            lottos.add(new Lotto(Utils.generateRandomLottoNumbers()));
+            lottos.add(new Lotto(Utils.generateRandomLottoNumbers(MINIMUM_LOTTO_NUMBER, MAXIMUM_LOTTO_NUMBER, NUMBER_OF_LOTTO)));
         }
         return lottos;
     }
